Enable explicit data transfers to gpu
