+++ compiler/rustc_span/src/symbol.rs
@@ -1588,6 +1588,10 @@ symbols! {
         prefetch_write_instruction,
         prefix_nops,
         preg,
+        preload,
+        preload_mut,
+        preload_type,
+        preload_mut_type,
         prelude,
         prelude_import,
         preserves_flags,
